[SCB-2582]bump apache version from 21 to 25
